Default Speed, Reload Config, Spigot Plugin
You can now set a default speed that the minecarts should go.

Added a /FastMinecarts Reload Command that uses the permission: "FastMinecarts.reload" but it can also be ran by an op.

It now runs on the spigot API because there was no need for the paper one and gives more compatibility for other plugins